rename articleDetails folder to ArticleModel
