+++ src/components/ArticlePageContent/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ArticleModel from '../../components/ArticleDetails';
+import ArticleModel from '../../components/ArticleModel';
 import { ArticlesSearch } from '../../components/ArticlesSearch';
 import { RecentArticlesContainer } from '../../components/RecentArticles';
 import { useCharityAPI } from '../../clients/index';
